pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel>=0.21",
    "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "Auto-FOX-Data"
dynamic = [
    "version",
    "readme",
]
description = "A repository with MD data used by the Auto-FOX tests."
license = { file = "LICENSE" }
authors = [
    { name = "Bas van Beek", email = "b.f.van.beek@vu.nl" },
]
keywords = [
    "science",
    "chemistry",
    "python-3",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Typing :: Typed",
]
requires-python = ">=3.7"
dependencies = [
]

[project.urls]
Homepage = "https://github.com/nlesc-nano/auto-FOX-data"
"Bug Tracker" = "https://github.com/nlesc-nano/auto-FOX-data/issues"

[project.optional-dependencies]
test = [
    "pytest>=6.0",
    "h5py",
    "numpy>=1.21",
    "assertionlib>=2.2.0",
    "pyyaml>=5.1",
    "Nano-Utils>=2.0",
    "ipython>=5.0.0",
]
lint = [
    "flake8",
    "pydocstyle[toml]>=6.1",
    "Flake8-pyproject>=1.2.2",
    "mypy",
    "types-PyYAML",
]

[tool.setuptools]
license-files = ["LICENSE*.txt"]

[tool.setuptools.packages.find]
include = ["FOXdata*"]

[tool.setuptools.package-data]
FOX = [
    "armc/*.hdf5",
    "armc/*/*.xyz",
    "armc/*/*.dill",
    "armcpt/*.hdf5",
    "armcpt/*/*.xyz",
    "armcpt/*/*.dill",
    "py.typed",
]

[tool.setuptools.dynamic]
readme = { file = ["README.rst"], content-type = "text/x-rst" }

[tool.setuptools_scm]
write_to = "FOXdata/_version.py"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
addopts = "--tb=short --pdbcls=IPython.terminal.debugger:TerminalPdb"

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "h5py.*",
]
ignore_missing_imports = true

[tool.pydocstyle]
convention = "numpy"

[tool.flake8]
max-line-length = 100
